ADE-10: Fixing email error
